Fix up VS Code settings for AmigaOS 4.x
Provide more appropriate paths for the 4.x SDK and for MUI, based on
default drawer layout for install on AmigaOS 4.x itself. MUI developer
files on 4.x are installed in a "MUI" drawer within the SDK installation
instead of in a "Developer" drawer within the MUI installation, unless
the user manually installs and configures MUI. SDK installs themselves
replace older versions of the 4.x SDK, so there's no version number in
the drawer name.