Set catalog descriptor properties for flexcat
One of the problems with loading the catalog turns out to be that
flexcat defaults to the value of -1 for version when calling
OpenCatalog(). This fails since the version tag value requires identical
version from the catalog file when the tag value is not 0.

Since the strings in iGame.cd are all English, we might as well state
that as well instead of relying on flexcat again giving us a default
value in the generated source code.