Move catalog rules into make_includes/catalogs.inc
With the exception of the call to flexcat, the catalog rules should be
identical for each makefile. Because we can't guarantee that the shell OR
operator will be present or work on Amiga OS or Windows, the makefiles for
these two platforms need to be able to have their own definitions of the
catalogs/%/iGame.catalog rule which do not include the "|| exit 0" at the
end.